requirements.txt
asgiref==3.4.1
break==0.0.1
coverage==6.0.2
Django==3.2.8
djangorestframework==3.12.4
ele==0.2.0
prn==0.84
pytz==2021.3
sqlparse==0.4.2
